refactor(monitors): import MonitorForm directly, drop index barrel
