--- web/src/features/monitors/index.ts
@@ -1,1 +1,0 @@
-export { MonitorForm } from "./components/MonitorForm";
