+++ web/src/features/monitors/pages/EditMonitorPage.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 
 import { ErrorPage } from "@/src/components/error-page";
 import Page from "@/src/components/layouts/page";
-import { MonitorForm } from "@/src/features/monitors";
+import { MonitorForm } from "@/src/features/monitors/components/MonitorForm";
 import { MonitorPagePermissions } from "@/src/features/monitors/components/MonitorPagePermissions";
 import { api } from "@/src/utils/api";
